# docs/openapi/swagger.yaml
components:
  schemas:
...
    ? github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-array_github_com_Mininglamp-OSS_octo-marketplace_internal_model_CategoryFilter
    : properties:
        data:
          items:
            $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_model.CategoryFilter'
          type: array
          uniqueItems: false
      type: object
    ? github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-array_github_com_Mininglamp-OSS_octo-marketplace_internal_model_TagFilter
    : properties:
        data:
          items:
            $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_model.TagFilter'
          type: array
          uniqueItems: false
      type: object
...
    ? github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_service_IconResult
    : properties:
        data:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_service.IconResult'
      type: object
    ? github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_service_ProbeResponse
    : properties:
        data:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_service.ProbeResponse'
      type: object
    ? github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_service_parse_IconUploadResult
    : properties:
        data:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_service_parse.IconUploadResult'
      type: object
...
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-internal_api_handler_expert_InstallExpertResp:
      properties:
        data:
          $ref: '#/components/schemas/internal_api_handler_expert.InstallExpertResp'
      type: object
...
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error:
      properties:
        error:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.ErrorBody'
      type: object
...
    github_com_Mininglamp-OSS_octo-marketplace_internal_service.ProbeRequest:
      properties:
        args:
          items:
            type: string
          type: array
          uniqueItems: false
        command:
          type: string
        env:
          additionalProperties:
            type: string
          type: object
        headers:
          additionalProperties:
            type: string
          type: object
        transport:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_model.Transport'
        url:
          type: string
      type: object
...
    internal_api_handler.MCPIconUploadRequest:
      properties:
        content_type:
          type: string
        file_name:
          type: string
        file_size:
          type: integer
      required:
      - file_name
      - file_size
      type: object
...
    internal_api_handler_expert.InstallExpertReq:
      properties:
        runtime_id:
          type: string
        workspace_id:
          type: string
      type: object
...
  securitySchemes:
    AdminToken:
      in: header
      name: X-Admin-Token
      type: apiKey
    Bearer:
      scheme: bearer
      type: http
info:
  contact:
    name: OCTO API Team
    url: https://github.com/Mininglamp-OSS/octo-marketplace
  description: Skill and MCP marketplace API for OCTO.
  title: Octo Marketplace API
  version: 1.0.0
openapi: 3.1.0
paths:
  /admin/mcps/_probe:
    post:
      description: Probe a remote MCP connection through the administrator surface
        without persisting it.
      operationId: admin_mcp.probe
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_service.ProbeRequest'
                description: Connection to probe
                summary: body
        description: Connection to probe
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_service_ProbeResponse'
          description: OK
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: VALIDATION_ERROR
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - AdminToken: []
      summary: Probe system MCP server
      tags:
      - admin_mcp
...
  /connectors/_probe:
    post:
      description: Probe a Connector's remote MCP configuration without persisting
        it; the legacy MCP probe remains available during rollout.
      operationId: connector.probe
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_service.ProbeRequest'
                description: Connector configuration to probe
                summary: body
        description: Connector configuration to probe
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_service_ProbeResponse'
          description: OK
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: VALIDATION_ERROR
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - Bearer: []
      summary: Probe connector
      tags:
      - plugin
...
  /experts/{expert_id}/install:
    post:
      description: Provision the expert as a Loop agent (with its skills) in the chosen
        workspace/runtime. Aggregates octo-fleet calls on behalf of the caller and
        rolls back on partial failure.
      operationId: expert.install
      parameters:
      - description: Expert ID
        in: path
        name: expert_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_api_handler_expert.InstallExpertReq'
                description: Target workspace + runtime
                summary: body
        description: Target workspace + runtime
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-internal_api_handler_expert_InstallExpertResp'
          description: OK
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: VALIDATION_ERROR
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: CONFLICT
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: UPSTREAM_UNAVAILABLE
      security:
      - Bearer: []
      summary: Install expert to a Loop workspace/runtime
      tags:
      - expert
...
  /mcp_categories:
    get:
      description: Return MCP category keys and visible record counts for the current
        Space.
      operationId: mcp_category.list
      parameters:
      - description: 'Scope: mine restricts counts to caller-owned records'
        in: query
        name: mode
        schema:
          type: string
      - description: Provenance filter (human, bot, import); repeatable or comma-separated
        in: query
        name: created_by_type
        schema:
          items:
            type: string
          type: array
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-array_github_com_Mininglamp-OSS_octo-marketplace_internal_model_CategoryFilter'
          description: OK
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - Bearer: []
      summary: List MCP categories
      tags:
      - mcp
  /mcp_icon_uploads:
    post:
      description: Create a presigned upload target and persistent URL for an MCP
        icon.
      operationId: mcp_icon_upload.create
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_api_handler.MCPIconUploadRequest'
                description: Icon metadata
                summary: body
        description: Icon metadata
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_service_parse_IconUploadResult'
          description: OK
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: VALIDATION_ERROR
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: PAYLOAD_TOO_LARGE
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - Bearer: []
      summary: Initialize MCP icon upload
      tags:
      - mcp
  /mcp_tags:
    get:
      description: Return tags aggregated from MCPs visible to the caller in the current
        Space. Free-form strings — no dedicated tag catalog table. Supports fuzzy
        match by `q` and sorts by descending row count (ties broken alphabetically).
      operationId: mcp_tag.list
      parameters:
      - description: Fuzzy tag search (case-insensitive substring)
        in: query
        name: q
        schema:
          type: string
      - description: Max items returned; default 50, max 100
        in: query
        name: limit
        schema:
          type: integer
      - description: 'Scope: mine restricts aggregation to caller-owned records (mirrors
          GET /mcps/mine)'
        in: query
        name: mode
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-array_github_com_Mininglamp-OSS_octo-marketplace_internal_model_TagFilter'
          description: OK
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - Bearer: []
      summary: List MCP tags
      tags:
      - mcp
...
  /mcps/_probe:
    post:
      description: Probe a remote MCP server connection without persisting the supplied
        configuration.
      operationId: mcp.probe
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_service.ProbeRequest'
                description: Connection to probe
                summary: body
        description: Connection to probe
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_service_ProbeResponse'
          description: OK
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: VALIDATION_ERROR
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - Bearer: []
      summary: Probe MCP server
      tags:
      - mcp
...
  /mcps/{mcp_id}/icon:
    post:
      description: Upload and replace the icon for an MCP server owned by the caller.
      operationId: mcp.icon.upload
      parameters:
      - description: MCP ID
        in: path
        name: mcp_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              format: binary
              title: file
              type: string
          multipart/form-data:
            schema:
              type: object
        description: Icon file
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_service_IconResult'
          description: OK
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: VALIDATION_ERROR
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: PAYLOAD_TOO_LARGE
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - Bearer: []
      summary: Upload MCP icon
      tags:
      - mcp
...
servers:
- url: /v1
...